# lib/api-zod/src/generated/types/accessState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Floor Ops Console API
 * OpenAPI spec version: 0.1.0
 */

export type AccessState = typeof AccessState[keyof typeof AccessState];


export const AccessState = {
  submitted: 'submitted',
  access_ok: 'access_ok',
  invite_pending: 'invite_pending',
  join_link: 'join_link',
  public_repl: 'public_repl',
  preview_only: 'preview_only',
  preview_stale: 'preview_stale',
  denied: 'denied',
} as const;