# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Floor Ops Console API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalTeams: number;
  pageNowCount: number;
  stuckCount: number;
  silentCount: number;
  doNotDisturbCount: number;
  draftTicketsCount: number;
  filedTicketsCount: number;
  resolvedTicketsCount: number;
}