# lib/api-zod/src/generated/types/eventSettingsUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Floor Ops Console API
 * OpenAPI spec version: 0.1.0
 */

export interface EventSettingsUpdate {
  /** @nullable */
  endsAt: Date | null;
}