# lib/api-zod/src/generated/types/helpType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Floor Ops Console API
 * OpenAPI spec version: 0.1.0
 */

export type HelpType = typeof HelpType[keyof typeof HelpType];


export const HelpType = {
  fine: 'fine',
  stuck: 'stuck',
  page_now: 'page_now',
  do_not_disturb: 'do_not_disturb',
} as const;