# lib/api-zod/src/generated/types/pagerQueueItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Floor Ops Console API
 * OpenAPI spec version: 0.1.0
 */
import type { HelpType } from './helpType';

export interface PagerQueueItem {
  /** @nullable */
  ticketId: number | null;
  teamId: number;
  teamName: string;
  tableLabel: string;
  blocker: string;
  helpType: HelpType;
  ageMinutes: number;
  moderatorEnRoute: boolean;
}