# lib/api-zod/src/generated/types/teamDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Floor Ops Console API
 * OpenAPI spec version: 0.1.0
 */
import type { Ping } from './ping';
import type { Team } from './team';
import type { Ticket } from './ticket';

export interface TeamDetail {
  team: Team;
  pings: Ping[];
  tickets: Ticket[];
}