# lib/api-zod/src/generated/types/ticketStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Floor Ops Console API
 * OpenAPI spec version: 0.1.0
 */

export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export const TicketStatus = {
  draft: 'draft',
  filed: 'filed',
  discarded: 'discarded',
  resolved: 'resolved',
} as const;